A restructuring of the way the vocabulary table is edited. The old Inline editing option is removed. You can now always type directly in the table. The Entry dialog can be invoked with a keyboard shortcut.
KVocTrain has always been designed to append new entries to the end of the list and will remain so. Smart appending works well for appending new entries using the dialog. Without the dialog the Insert key will append a new row as usual. Minor improvements can probably still be made but this should be enough to close bugs 93172, 94318, and 95639.

BUG: 93172
BUG: 94318
BUG: 95639

svn path=/trunk/kdeedu/kvoctrain/; revision=376338