ALL preferences are know ported to KConfigXT. Next step is to also make a managed settings dialog. This changes the way some settings are stored. For the user this means that:
-All languages must be redefined
-All stored settings profiles must be redefined

However, this is done easily enough so I'm not going to bother with conversion routines. Some things may change again with the managed dialog.

I would appreciate some testing of this to make sure all preferences are stored and retrieved as expected. This commit also takes care of wish #99300.

BUG:99300
CCMAIL:kde-edu@kde.org

svn path=/trunk/kdeedu/kvoctrain/; revision=399065