GUI: A new configuration dialog for KVocTrain.
This is a new KConfigXT configuration dialog. It combines the two old dialogs Configure KVocTrain and Configure Query.

The new code is in kvoctrain/kvoctrain/common-dialogs.

The old code in kvoctrain/kvoctrain/option-dialogs and kvoctrain/kvoctrain/queryoption-dialogs is still being compiled, but is not used. I will remove it after some further testing.

svn path=/trunk/kdeedu/kvoctrain/; revision=408920