This is the initial commit of a possible design for the new "EduWidget" Parley practice dialog framework (a possible gsoc project).
At the moment, it is only headers; this is because it is for design purposes.

The general overview of the proposed framework is as follows:
There would be a set of EduWidgets, for lack of a better name, that would be used to create the various practice dialogs.
Each EduWidget represents a particular common element in a practice mode (question display, user input, statistics, etc), and it would
cotain all the logic to do its task within it. Each practice mode would use the QWidget that makes sense for its purposes to display
the EduWidget (for example, the user input would be a LineEdit in the question & answer mode). 

Using the signal/slots mechanism, making a new mode would be as easy as selecting which EduWidgets you want, picking how you want
to display them, and adding them to an SVG template.

For more detail, look at my gsoc application [me = David Capel] ;) Send questions/comments to wot.narg@gmail.com


svn path=/trunk/KDE/kdeedu/parley/; revision=797473