This commit is a major redesign that was suggested by fregl. The entry-related logic has been taken out of the Input and Prompt classes and moved into the TestEntryManager and AnswerValidator classes. Additionally, some files from the old practice directory have been moved here and subsumed into this project.
Notable changes from the previous TestEntryManager and AnswerValidator classes are the elimination of the TestEntry class (it is redundant in my design and 
will only get in the way) and the use of signals instead of return values, which fits my design of very loosely-fitting practice elements.

Additionally, most common actions have been changed into KActions and the prefs.h in from the parley src has been copied and changed to practiceprefs.h (and 
reduced to only relavent config options).

This commit compiles but crashes on startup. I'll hunt down the crash bug later, but I want to get this change commited.



svn path=/branches/work/soc-parley/parley/; revision=814012