Using svgs as layouts now works. Layouts are in layout.svgz and independent widgets are in widgets.svgz. Once things stabilize a bit I'll draw up documentation for those files (for themeing use).
The svg percent bar works -- it is the first example of the upcoming themeing system.

Known bug: when resizing the window, the svg resizes to match the previous window size instead of the current one. Something is interacting wierdly with the
qt function I'm using...

TODO: more svg stuff. I need to make (or have made ;) ) a more complex layout svg to work with.
      change the prompt to svg text instead of a label?


svn path=/branches/work/soc-parley/parley/; revision=817084