Major remodeling of how themes work, based on Frederik's idea.
Now there is a group named 'main' that is the main background, and it has a rectangle named 'active_area' to denote where to put the rest.

For each mode, the themer makes a group named that name (written => 'written', etc). Within that group are all the requisit elements (answer_text_input, etc).
Depending on the mode, practice puts the named group in the 'actve_area' rectangle, and we are off! For certain modes, there is another mode that is used as 
a fallback. For example, if no 'example' element is found, it uses the 'written' element instead (they are similar enough most themes just won't both making a 
seperate design for example.) This will be documented more thoroughly later.

This change allows for more freedom when designing themes. 

For now, don't try to use any modes except for written, example, and paraphrase.


svn path=/trunk/KDE/kdeedu/parley/; revision=840889