Addition of Comparison mode and related infrastructure.
This involved a fair amount of infrastructure becaused it requires multiple inputs and multiple outputs.
Adapting the various signals/slots was easy, but making a sane widget manager was more difficult. I settled on
an implementation that you'd pass a list of elementIds and a 'Logic' class to, and it would manage the GUI details
and forward all the decisions to the logic class. This will make it reusable, since new modes only need to reimplement
the MWPLogic interface.

Speaking of new modes, comparison was somewhat of a test run -- conjugation is next, and it will be the real challenge.

svn path=/trunk/KDE/kdeedu/parley/; revision=842938