begin refactoring the Parley main window
The goal is to have a common main window and components for the editor, the practice window and the
welcome screen that are merged into the main window with xmlgui.

The first step was to rename ParleyApp to ParleyMainWindow, I also copied the ParleyApp code to a new
class called Editor. The next step will be to separate main window code from editor code.

svn path=/trunk/KDE/kdeedu/parley/; revision=878718