Use deleteLater() instead of delete when deleting old components.
This fixes a crash when doubleclicking on an item in the welcome screen.
WelcomeScreen::slotDoubleClicked() opened the document and switched to the
editor, which caused the WelcomeScreen to be deleted immediately. Apparently
the (deleted) QListView tries to do some more stuff after emitting doubleClicked(),
which isn't really the best thing to do. ;)

svn path=/branches/work/parley-practice/parley/; revision=1093000