Don't crash when downloading .tar.7z files with KNS.
If I just download it manually, uncompress the *.tar.7z file and feed the kvtml
file into parley it doesn't crash.

There are 2 separate problems involved here:
1) KNS doesn't work with *.tar.7z files.
2) Even if the the file format of the file downloaded from KNS isn't recognized
   for now, parley shouldn't crash.

The patch in this reviewboard is intended to fix the problem 2) and it was made
with help from Aleix Pol (thanks!).

Just in case you are curious about the problem 1) - the related code from
kdelibs is the following:

* knewstuff/knewstuff3/core/installation.cpp
  (function Installation::installDownloadedFileAndUncompress)
  Here knewstuff tries to use KTar to deal with the downloaded file.

* files under kdecore/compression/*
  Here is kfilterdev, if I'm not mistaken this is what KTar uses to deal with
  files compressed gz,bzip,etc. The thing is: there is no filter for *.7z files,
  so that's why they can't work in KNS right now.

REVIEW: 113484