Removes gradient in grade statistics.
This closes the mentioned bugs, as it removes the gradient in learning statistics and uses solid color.
It also makes it even more intuitive for the user, as it uses a single color for all the grades with different alpha values, increasing as grades increase. Thus darker shades mean higher grades.
Also, the grade color settings in Parley configuration have been changed, removing the different color palletes for all grades and leaving just a single base color for all the grades.
The "Use colors" option has been removed from the configuration.
Also, the font color for the translations in the editor has been changed to be the  default black, making the text more readable.

BUG: #266368 , #328279

REVIEW: 115449