Add default unittest setup to parley
This creates 2 stub unittests in the autotest directory so that devs
can easily add new unit tests of classes without doing the setup in
this patch.

It moves the definition of parley's include directories and
link libraries up into the root CMake file, so that they are visible
to the unittests and only need to be redefined in one location when a
dev makes a change elsewhere.

REVIEW: 118895