Fix null pointer reference in SynonymWidget::togglePair() by initializing pointer
In both cases (antonyms and falseFriends) the synonymWidget::togglePair() was referencing an uninitialized pointer m_doc to call setModified() on the  KEduVocDocument.  The quick fix was to store pointers to the antonym and falseFriend widgets and initialize m_doc.  A better long term fix would be to have the kedevocdoc manage its own state and determine when it is dirty on its own.

BUG: 336419
BUG: 336422

REVIEW: 119107