Changed rename of timed backup file to copy in parleydocument.cpp
Parley was renaming the timed backup just before trying to save.
If it didn't recognize the file format then it would fail the save,
but not restore the backup.
By copying the backup kvocdoc overwrites the timed backup file and even
when the file format is unrecognized the fallback is make a copy
in filename~ and do nothing to filename

I added todo note that timed backup should be handled by kvocdoc
which is already handling open, saving, locking and closing the file.

BUG: 336820