Put BarWidget into its own file.
This follows the kdelibs coding standard better.

I also had to create a collectionwidget class to host some variables but that
is not yet implemented. This will later become the widget class that shows the
overview of the collection. This is now created on the fly inside the
dashboard, a.k.a. the welcomescreen.