Fix bugs: 338302 and 338417 title bar [modified] tag is inconsistent
Symptoms:
The [modified] in the title bar was inconsistent.

Problems:
1. The title bar was only connected to the dirty bit in the doc from
the Editor. Going directly welcome -> stats did not connect the title
bar to the dirty bit.
2.  Whenever a new document was created, new collection, welcome screen
etc. the title bar was disconnected.
3. In several places setModified was called on the kvocdoc when either
nothing changed or the kvocdoc should know better.

Solution:
The title bar is ParleyMainWindow's title bar. ParleyMainWindow also
know when the document is updated. I created a updateParleyDocument
so that ParleyMainWindow always knows when the doc changes and
slotFileNew so that anyone can ask ParleyMainWindow to update its
own title bar.

Not fixed: Parley still directly calls the setModified in kvocdoc
when it starts practice. kvocdoc needs to be fixed before that is
removed.

BUG:338302
BUG:338417
FIXED-IN:4.14.1