Prevent array access with negativ index
To adjust the practice date of expired entries
in EntryFilter::expireEntries() the expire item
time was read from the Prefs object. When grade
of the expired item was 1, the argument of
Prefs::expireItem() became -1. This resulted in
an access of an C array with negativ index in
Prefs::expireItem(). For grade 2 the prefs value
was not defined in the config file - zero was
returned by default. For this two cases the
practice date was set to current time as a
defined state.

REVIEW: 127157