Fix EntryFilter::expireEntries call.
In EntryFilter::expireEntries() the set
EntryFilter::m_entries is checked for entries
which hasn't practiced for a defined amount of
time and decreases the grad if this was expired.

But at time of calling expireEntries() this could
not work because m_entries was not intialisised
and empty. The method call was moved to a place
where m_entries is filled and cleaned of invalid
entries.

REVIEW: 127156