Fix never ending event loop when quitting app and segmentation faults
When the app is quited app.exec() doesn't return. The process stays in
sleeping state. Removed app.setQuitOnLastWindowClosed(false) because
the quit() or exec() method of the application were not invoked
somewhere else in the application.

After this the app ends in the correct way. But after quit is invoked,
there are sporadic segmentation faults and the following message was
shown on stderr:
    QCoreApplication::postEvent: Unexpected null receiver
This was fixed by removing zero pointer access in destuctor of
ParleyDocument.

Added missing this pointer on KEduVocDocument allocation.

REVIEW: 127441