Fix Coverty Scan ID 1342699 - Uninitialized pointer field
One integer member m_numSets was initilized with 0 and two pointer
members were removed by refactoring the EntryFilter class.

The code for a dialog window (where the two former pointers was used
for) was moved into a new created class EntyFilterDialog. The dialog
is now created as a stack object and will be removed properly when
goes out of scope.

REVIEW: 127558