Fix Coverty Scan ID 1342595 - Dereference before null check
Refactoring of class Collection. The pointer access before null check
is located in the close() method wich is called in the destructor.
Further invetigation showed that this call is useless because the
KEduVocDocument is opened read only (see LibKEduVocDocument::close()).
The close method call was removed.

While further refactoring more unused fuctionality was removed.

REVIEW: 127658