Fix Coverty Scan ID 1343237 - Unused value
The coverty scan issue in BlockOptions::checkValidity() was fixed by
moving the boolean variable "found" outside the for loop. By this it
wouldn't reset to false at the beginning of each loop cycle and works
now as expected.

By the way BlockOptions::checkValidity() was refactored to run the for
loops only when checkboxes are enabled. Also a validy check was added
when check boxes becomes checked. Removed unused QString variable.

REVIEW: 127663