Fix theme cache path handling
When path to theme cache file does not exitst (e.g. after a fresh
installation) it is not created automatically which caused the
following error to be shown on stderr while reading and writing:

"QIODevice::read (QFile, "/path/to/theme/cache/startpagethemecache.bin"
): device not open"

Automatic creation of theme cache path is added. Location of cache file
is moved to QStandardPaths::CacheLocation.

REVIEW: 127881