Add locale name to column title
In editor mode, when the same column is displayed, coming from different languages, it is impossible to find out the source language. E.g. when Word Type is visible, for both English and Spanish, both columns are displayed as Word Type, at the header of the table. In terms of this solution, locale name has been added to each column title (since it is more compact than language name).

REVIEW: 130124