[Parley] Fix false error message when entering dashboard
Summary:
When entering dashboard with conjugations as current practice mode, an
error message box is shown when no entries are due for practice. It
says: "The vocabulary document contains no entries that can be used for
the chosen type of practice."
For determining the word count of the collection widgets, the
EntryFilter class is used. A missing test for the showDialg flag was
added in the conjugation entries evaluation.

BUG: 387656

Reviewers: #kde_edu, apol

Reviewed By: apol

Subscribers: apol

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D9246