Fix segfault when leaving statistics main window
Reproducible segfault when leaving statistics main window with unsaved
changes and clicking on save in the query dialog.
The segfault is thrown by KEduVocDocument pointer access which was
deleted before during destruction of StatisticsMainWindow. Both document
and widget are destructed by deleteLater. With the conditions mentioned
above the order of destruction becomes to the document deleted first
which leads to the segfault.
The fix is done by changing the pointer for KEduVocDocument to a shared
pointer which was not deleted until the last using class is deleted.

BUG: 420302