Fix assert with Qt debug build in QWidget destructor
This fixes an assert in the QWidget destructor with message
"d->focus_next->d_func()->focus_prev == this" upon calling the
destructor of LanguagePropertiesPage.
The reason is that the page is destroyed too early via a parent
connection, before the KPageWidgetItem, which owns the page,
destroys it and handles the focus correctly. Removing the parent
does not lead to a memory leak because KPageWidgetItem takes care
of the destruction.